# perf: Fix memory leak in cachedIterator