--- a/pkg/iter/cache.go
+++ b/pkg/iter/cache.go
@@ -98,6 +98,9 @@ func (it *cachedIterator) Err() error { return it.iterErr }
 
 func (it *cachedIterator) Close() error {
 	it.Reset()
+	if it.wrapped != nil {
+		it.wrapped.Close()
+	}
 	return it.closeErr
 }
 
@@ -194,5 +197,8 @@ func (it *cachedSampleIterator) Err() error { return it.iterErr }
 
 func (it *cachedSampleIterator) Close() error {
 	it.Reset()
+	if it.wrapped != nil {
+		it.wrapped.Close()
+	}
 	return it.closeErr
 }